Revert v1beta1 endpoint to v1alpha1, adding the new RSQL capabilities
