+++ cli/src/test/resources/mappings/sbomer/sbom-123-search.json
@@ -1,7 +1,7 @@
 {
   "request": {
     "method": "GET",
-    "url": "/api/v1beta1/sboms?pageSize=1&pageIndex=0&query=id%3D%3D123"
+    "url": "/api/v1alpha1/sboms?pageSize=1&pageIndex=0&query=id%3D%3D123"
   },
   "response": {
     "status": 200,
