+++ cli/src/test/resources/mappings/sbomer/sbom-123-to-process.json
@@ -1,7 +1,7 @@
 {
   "request": {
     "method": "GET",
-    "url": "/api/v1beta1/sboms/123-to-process"
+    "url": "/api/v1alpha1/sboms/123-to-process"
   },
   "response": {
     "status": 200,
