+++ cli/src/test/resources/mappings/sbomer/sbom-123.json
@@ -1,7 +1,7 @@
 {
   "request": {
     "method": "GET",
-    "url": "/api/v1beta1/sboms/123"
+    "url": "/api/v1alpha1/sboms/123"
   },
   "response": {
     "status": 200,
