+++ service/src/main/resources/application.yaml
@@ -36,7 +36,7 @@ quarkus:
         level: DEBUG
 
 sbomer:
-  api-url: "https://${SBOMER_ROUTE_HOST}/api/v1beta1/"
+  api-url: "https://${SBOMER_ROUTE_HOST}/api/v1alpha1/"
   # See documentation about features and feature flags
   features:
     umb:
@@ -99,7 +99,7 @@ sbomer:
       password: ""
 
   sbomer:
-    api-url: "http://localhost:8080/api/v1beta1/"
+    api-url: "http://localhost:8080/api/v1alpha1/"
     features:
       umb:
         enabled: false
@@ -125,7 +125,7 @@ sbomer:
       password: ""
 
   sbomer:
-    api-url: "http://localhost:8080/api/v1beta1/"
+    api-url: "http://localhost:8080/api/v1alpha1/"
     features:
       umb:
         enabled: true
