+++ service/src/test/java/org/jboss/sbomer/test/messaging/GenerationFinishedMessageBodyTest.java
@@ -64,12 +64,12 @@ public class GenerationFinishedMessageBodyTest {
                 .sbom(
                         Sbom.builder()
                                 .id("429305915731435500")
-                                .link("https://sbomer/api/v1beta1/sboms/429305915731435500")
+                                .link("https://sbomer/api/v1alpha1/sboms/429305915731435500")
                                 .bom(
                                         Bom.builder()
                                                 .format(BomFormat.CYCLONEDX)
                                                 .version("1.4")
-                                                .link("https://sbomer/api/v1beta1/sboms/429305915731435500/bom")
+                                                .link("https://sbomer/api/v1alpha1/sboms/429305915731435500/bom")
                                                 .build())
                                 .build())
                 .build(
